refactor: 응답으로 dto 전달하도록 controller, service, repo 수정

+++ src/main/java/com/sprint/mission/discodeit/service/basic/BasicBinaryContentService.java
@@ -1,7 +1,9 @@
 package com.sprint.mission.discodeit.service.basic;
 
+import com.sprint.mission.discodeit.dto.data.BinaryContentDto;
 import com.sprint.mission.discodeit.dto.request.binarycontent.BinaryContentCreateRequest;
 import com.sprint.mission.discodeit.entity.BinaryContent;
+import com.sprint.mission.discodeit.mapper.BinaryContentMapper;
 import com.sprint.mission.discodeit.repository.BinaryContentRepository;
 import com.sprint.mission.discodeit.service.BinaryContentService;
 import lombok.RequiredArgsConstructor;
@@ -16,33 +18,37 @@ import java.util.UUID;
 @RequiredArgsConstructor
 public class BasicBinaryContentService implements BinaryContentService {
     private final BinaryContentRepository binaryContentRepository;
+    private final BinaryContentMapper binaryContentMapper;
 
     @Transactional
     @Override
-    public BinaryContent create(BinaryContentCreateRequest request) {
+    public BinaryContentDto create(BinaryContentCreateRequest request) {
         String fileName = request.fileName();
         String contentType = request.contentType();
         byte[] bytes = request.bytes();
         BinaryContent binaryContent = new BinaryContent(fileName, (long) bytes.length, contentType, bytes);
-        return binaryContentRepository.save(binaryContent);
+        return binaryContentMapper.toDto(binaryContentRepository.save(binaryContent));
     }
 
     @Override
-    public BinaryContent find(UUID id) {
-        return binaryContentRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("BinaryContent not found: " + id));
+    public BinaryContentDto find(UUID id) {
+        return binaryContentMapper.toDto(binaryContentRepository.findById(id)
+                .orElseThrow(() -> new NoSuchElementException("BinaryContent not found: " + id)));
     }
 
     @Override
-    public List<BinaryContent> findAllByIdIn(List<UUID> ids) {
-        return binaryContentRepository.findAllByIdIn(ids);
+    public List<BinaryContentDto> findAllByIdIn(List<UUID> ids) {
+        return binaryContentRepository.findAllByIdIn(ids).stream()
+                .map(binaryContentMapper::toDto)
+                .toList();
     }
 
     @Transactional
     @Override
     public void delete(UUID id) {
-        binaryContentRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("BinaryContent not found: " + id));
+        if (!binaryContentRepository.existsById(id)) {
+            throw new NoSuchElementException("BinaryContent not found: " + id);
+        }
 
         binaryContentRepository.deleteById(id);
     }
